feat/transmissibility: fix labels and drop regularized logistic regressions
Logistic regression with too few data run into perfect separation
problems, which forced me to add fake data (ones and zeros everywhere)
to get the graphs to build. This is all good in some cases, but in most
of our cases it really made the graphs all wrong, faking an evolution in
transmissibility that did not exist. So we're sticking to averages and
no logistic regression for these graphs.