feat/model: use a consensus alignment for all plots
This fixes what was an undecided issue up to now, namely how to deal
with the multiplicity of deep alignments. One solution was to take all
DAs into account and weigh accordingly, which works for feature plots
but not for the latest plots we've been doing in the model exploration.
So we rationalised the approach to plotting lineag tree, just a little
to make it consistent and have it produce valid consensus alignments.
This is now the DA used for all plots in model exploration, and future
commits should propagate this to other notebooks.

This work also fixes a bug that was found in the chunk indentification
code; propagating this consensus algorithm to other notebooks will be
enough to fix them.